Migrating all of the helpers into the helpers file
